+++ dpc-attribution/src/main/java/gov/cms/dpc/attribution/AttributionServiceModule.java
@@ -1,12 +1,21 @@
 package gov.cms.dpc.attribution;
 
 import com.google.inject.Binder;
+import com.google.inject.Provides;
 import com.hubspot.dropwizard.guicier.DropwizardAwareModule;
 import gov.cms.dpc.attribution.service.LookBackService;
+import gov.cms.dpc.attribution.service.LookBackServiceImpl;
+import gov.cms.dpc.queue.service.DataService;
 
 public class AttributionServiceModule extends DropwizardAwareModule<DPCAttributionConfiguration> {
+
     @Override
     public void configure(Binder binder) {
-        binder.bind(LookBackService.class);
+    }
+
+    @Provides
+    LookBackService provideLookBackService(DataService dataService) {
+        return getConfiguration().isMockLookback() ? (orgId, patientID, providerID, withinMonth) -> true
+                : new LookBackServiceImpl(dataService);
     }
 }
