+++ dpc-attribution/src/main/resources/local.application.conf
@@ -21,4 +21,5 @@ dpc.attribution {
   }
 
   exportPath = "/app/data"
+  mockLookback = true
 }
