+++ dpc-attribution/src/test/java/gov/cms/dpc/attribution/TestServiceModule.java
@@ -1,6 +1,7 @@
 package gov.cms.dpc.attribution;
 
 import com.google.inject.Binder;
+import com.google.inject.Provides;
 import com.hubspot.dropwizard.guicier.DropwizardAwareModule;
 import gov.cms.dpc.attribution.service.LookBackService;
 
@@ -10,10 +11,13 @@ public class TestServiceModule extends DropwizardAwareModule<DPCAttributionConfi
     public TestServiceModule(LookBackService lookBackService) {
         this.lookBackService = lookBackService;
     }
+
     @Override
     public void configure(Binder binder) {
-        binder.bind(LookBackService.class).toInstance(lookBackService);
+    }
 
-
+    @Provides
+    LookBackService provideLookBackService() {
+        return lookBackService;
     }
 }
